Revert "Develop to Develop-Design."

+++ BataDase/MVVM/Models/MenuVMS/UsersM.cs
@@ -10,11 +10,11 @@ namespace BataDase.MVVM.Models.MenuVMS
         public string _surname { get; set; }
         public string _name { get; set; }
         public string _lastname { get; set; }
-        public int _sex { get; set; }
-        public double _birthday { get; set; }
+        public byte _sex { get; set; }
+        public long _birthday { get; set; }
         public string _login { get; set; }
         public string _pass { get; set; }
-        public int _access { get; set; }
+        public byte _access { get; set; }
         public float _bill { get; set; }
     }
 }
